requirements.txt
aiohttp>=3.9.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
redis>=5.0.0
cryptography>=42.0.0
